Crate zkp_common

zkp_common includes public constants G and H and data transfer objects (DTOs) used between client and server.

Constraints for G and H

Smaller values set for convenience when computing y1 = g^x and y2 = h^x

// public info
// Used when client is making the register commit
// r1 = g^k, r2 = h^k
// where k is a random seed generated by client
pub const G: u128 = 1;
pub const H: u128 = 2;

Two types of DTOs used:

For Client Requests

#[derive(Clone, Debug, Serialize, Deserialize)]
pub enum ClientRequest {
    Register(Username, Commits), // server will respond with a success or failure
    Authenticate(Username),      // server will respond with a challenge to this
    ProveAuthentication(Username, Answer), // server will respond with a sucess or failure
    CheckStatus(Username), // server will respond whether logged in or not. (called after Login flow)
}

For Server Responses

#[derive(Clone, Debug, Serialize, Deserialize)]
pub enum ServerResponse {
    Success,
    Challenge(u32),
    Failure(String),
}

Data Exchange Format

This crate also includes a helper function to write and flush data to stream.

For simplicity, JSON is used for data exchange

This can be replaced by a binary based representation such as Protocol Buffers for better performance

Modules

request_dto

response_dto

Constants

G

H

Functions

write_and_flush_stream